requirements.txt
absl-py==1.4.0
aiohttp==3.8.5
aiosignal==1.2.0
anyio==3.5.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.0.5
async-lru==2.0.4
async-timeout==4.0.2
attrs==23.1.0
Babel==2.11.0
backcall==0.2.0
beautifulsoup4==4.12.2
bleach==4.1.0
blinker==1.6.2
brotlipy==0.7.0
cachetools==4.2.2
certifi==2024.7.4
cffi==1.15.1
chardet==4.0.0
charset-normalizer==2.0.4
click==8.0.4
comm==0.1.2
conda-pack==0.7.1
contourpy==1.0.5
cryptography==41.0.3
cycler==0.11.0
ddsmoothing-python==1.0.0
debugpy==1.6.7
decorator==5.1.1
defusedxml==0.7.1
executing==0.8.3
fastjsonschema==2.16.2
fonttools==4.25.0
fqdn==1.5.1
frozenlist==1.3.3
google-auth==2.22.0
google-auth-oauthlib==0.5.2
grpcio==1.48.2
idna==2.10
importlib-metadata==6.0.0
importlib-resources==5.2.0
ipykernel==6.25.0
ipython==8.12.2
ipywidgets==8.0.4
isoduration==20.11.0
jedi==0.18.1
Jinja2==3.1.2
joblib==0.14.1
json5==0.9.6
jsonpointer==2.4
jsonschema==4.19.2
jsonschema-specifications==2023.7.1
jupyter==1.0.0
jupyter_client==8.6.0
jupyter-console==6.6.3
jupyter_core==5.5.0
jupyter-events==0.8.0
jupyter-lsp==2.2.0
jupyter_server==2.10.0
jupyter_server_terminals==0.4.4
jupyterlab==4.0.8
jupyterlab-pygments==0.1.2
jupyterlab_server==2.25.1
jupyterlab-widgets==3.0.9
kiwisolver==1.4.4
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.7.2
matplotlib-inline==0.1.6
mistune==2.0.4
mkl-fft==1.3.8
mkl-random==1.2.4
mkl-service==2.4.0
multidict==6.0.2
munkres==1.1.4
nbclient==0.8.0
nbconvert==7.10.0
nbformat==5.9.2
nest-asyncio==1.5.6
notebook==7.0.6
notebook_shim==0.2.3
numpy==1.24.4
oauthlib==3.2.2
opencv-python==4.11.0.86
overrides==7.4.0
packaging==23.1
pandas==2.0.3
pandocfilters==1.5.0
parso==0.8.3
patsy==0.5.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.0.0
pip==23.2.1
pkgutil_resolve_name==1.3.10
platformdirs==3.10.0
ply==3.11
prometheus-client==0.14.1
prompt-toolkit==3.0.36
protobuf==3.20.3
psutil==5.9.0
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
Pygments==2.15.1
PyJWT==2.4.0
pyOpenSSL==23.2.0
pyparsing==3.0.9
PyQt5-sip==12.11.0
PySocks==1.7.1
python-dateutil==2.8.2
python-json-logger==2.0.7
pytz==2023.3.post1
PyYAML==6.0.1
pyzmq==25.1.0
qtconsole==5.5.0
QtPy==2.4.1
referencing==0.30.2
requests==2.25.1
requests-oauthlib==1.3.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.10.6
rsa==4.7.2
scikit-learn==0.22.2.post1
scipy==1.10.1
Send2Trash==1.8.2
setuptools==68.0.0
sip==6.6.2
six==1.14.0
sklearn==0.0
sniffio==1.2.0
soupsieve==2.5
stack-data==0.2.0
statsmodels==0.14.0
tensorboard==2.12.1
tensorboard-data-server==0.7.0
tensorboard-plugin-wit==1.8.1
terminado==0.17.1
tinycss2==1.2.1
toml==0.10.2
tomli==2.0.1
torch==1.8.1+cu111
torchattacks==3.5.1
torchaudio==0.8.1
torchvision==0.9.1+cu111
tornado==6.3.3
tqdm==4.67.1
traitlets==5.7.1
ttach==0.0.3
types-python-dateutil==2.8.19.14
typing_extensions==4.7.1
tzdata==2023.3
uri-template==1.3.0
urllib3==1.26.16
wcwidth==0.2.5
webcolors==1.13
webencodings==0.5.1
websocket-client==0.58.0
Werkzeug==2.2.3
wheel==0.38.4
widgetsnbextension==4.0.5
yarl==1.8.1
zipp==3.11.0
